requirements.txt
# Core Libraries
numpy
pandas
scipy
matplotlib
scikit-learn

# Medical Image Processing
nibabel
pydicom
SimpleITK
medpy

# Deep Learning Frameworks
torch
torchvision
#tensorflow
keras

# Natural Language Processing
spacy
nltk
transformers

# Image Processing and Augmentation
opencv-python
albumentations

# Utilities and Development Tools
jupyterlab
ipykernel
tqdm
seaborn

# Experiment Tracking
mlflow

# Data Security
cryptography

# API Development
fastapi
uvicorn

# Deployment
docker

# Medical AI utilities
monai

# UI
streamlit
open-clip-torch
faiss-cpu
accelerate
sacremoses
google-generativeai